backend complete

+++ Backend/ExpenseMgmtApp_V1/src/main/java/com/itshaala/expensemgmtapp_v1/Dao/ExpenseDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExpenseDao extends JpaRepository<Expense, Integer> {
+public interface ExpenseDao extends JpaRepository<Expense, Long> {
 }
